fix(Columns): 컬럼의 psition 값 Long 타입으로 변경

+++ src/main/java/com/project/skilled_project/domain/columns/dto/request/ColumnsCreateRequestDto.java
@@ -7,6 +7,6 @@ public class ColumnsCreateRequestDto {
 
   private Long boardId;
   private String title;
-  private double position;
+  private Long position;
 
 }
