+++ src/main/java/com/project/skilled_project/domain/columns/repository/ColumnsRepositoryCustom.java
@@ -1,7 +1,7 @@
 package com.project.skilled_project.domain.columns.repository;
 
 public interface ColumnsRepositoryCustom {
-  double getCenterColumnsPositionTargetBig(Long boardId, double nowPosition, double targetPosition);
-  double getCenterColumnsPositionNowBig(Long boardId, double nowPosition, double targetPosition);
+  Long getCenterColumnsPositionTargetBig(Long boardId, Long nowPosition, Long targetPosition);
+  Long getCenterColumnsPositionNowBig(Long boardId, Long nowPosition, Long targetPosition);
 
 }
